resolve Ember data deprecation for testing tools #741
We have to stub the cardstack-data service in some tests
because it has async calls that the tests don't wait for
before destroying the store.

closes #741
